Updates shell script to log output somewhere and not run Lua scripts if there aren't any new emails.

+++ checkmail.sh
@@ -2,8 +2,10 @@
 
 getmail
 
-for f in /home/alex/Programming/smsweather/mail/new/*
-do
-	lua /home/alex/Programming/smsweather/smsweather.lua $f
-	rm $f
-done
+if [ "$(ls -A /home/alex/Programming/smsweather/mail/new)" ]; then
+	for f in /home/alex/Programming/smsweather/mail/new/*
+	do
+		lua /home/alex/Programming/smsweather/smsweather.lua $f &>>/home/alex/Programming/smsweather/luaerrs
+		rm $f
+	done
+fi
